Teams notification resource for concourse
Logging implemented in teams, using python to collect event stream of
logging from concourse. This is achieved by setting a watermark in the
log in the python script and collecting data from each event up to that point.
A curl is made to a teams webhook, that sends a message card with
information relating to the concourse, pipeline, job, task, and logs
from said task. We pass in atc_username, atc_password, webhook, proxy
and status to the resource which is used to create a message card to
send to teams.