MS Teams Concourse Resource
This modifies the forked code in order to add the ability to provide custom messages and a custom message file.
The title will still remain the same with the pipeline name and job name but the custom message is just below that with a
button that links to the concourse server.

Resolves VITE-149